Move ButlerFactory to a new file and modernize it slightly.
"Modernize" just means "switch to snake_case and use butler methods
instead of registry methods".
